codegen: cast the hidden-result register to the parameter's type
The Windows job failed to compile: an unfixed enum's underlying type is
unsigned under the Itanium ABI and signed under the MS one, so passing
the X64Register variable holding the hidden-result register to
codegen_canonical_x64_rsp_address's u32 parameter is a silent widening
on Linux and a -Wsign-conversion error on a host-Windows build. The
other three calls the wide-vector work added pass either an enum
constant or a u8, neither of which converts.

Reproduced locally by giving X64Register a signed fixed underlying type
and compiling the translation unit with -Wsign-conversion: the harness
reports the same diagnostic at the same column with the cast removed and
nothing at all with it in place, so this was the only site of its kind
in the file rather than the first of several.

+++ src/buster/lib/compiler/codegen/codegen.c
@@ -11308,7 +11308,12 @@ BUSTER_GLOBAL_LOCAL CodegenModule codegen_generate_canonical_module_attempt(Aren
                                 // reserved outgoing-area slot: the callee may
                                 // store it back with alignment-checking moves
                                 // the sixteen-aligned frame slot cannot take.
-                                if (!codegen_canonical_x64_rsp_address(&buffer, indirect_return_register, call_layout.windows_result_copy_offset,
+                                // The cast is not redundant: an unfixed enum is
+                                // unsigned under the Itanium ABI and signed
+                                // under the MS one, so a host-Windows build
+                                // sees a sign conversion here that a Linux one
+                                // does not.
+                                if (!codegen_canonical_x64_rsp_address(&buffer, (u32)indirect_return_register, call_layout.windows_result_copy_offset,
                                                                         call_layout.windows_result_copy_alignment))
                                 {
                                     result.error = buffer.error != CODEGEN_ERROR_NONE ? buffer.error : CODEGEN_ERROR_CAPACITY;
